Restructured patients component directory
