+++ src/components/Patients/EditPatient.jsx
@@ -1,7 +1,7 @@
 import React, {Component, Fragment} from 'react';
 import * as PropTypes from 'prop-types';
 import PatientForm from "./PatientForm";
-import PatientsService from "../../../services/PatientsService";
+import PatientsService from "../../services/PatientsService";
 import {withRouter} from "react-router-dom";
 import {withSnackbar} from "notistack";
