+++ src/components/Patients/NewPatient.jsx
@@ -1,8 +1,8 @@
 import React, {Component, Fragment} from 'react';
 import * as PropTypes from 'prop-types';
 import PatientForm from "./PatientForm";
-import Patient from "../../../models/Patient";
-import PatientsService from "../../../services/PatientsService";
+import Patient from "../../models/Patient";
+import PatientsService from "../../services/PatientsService";
 import {withRouter} from "react-router-dom";
 import {withSnackbar} from "notistack";
