+++ src/components/Patients/PatientForm.jsx
@@ -3,14 +3,14 @@ import * as PropTypes from 'prop-types';
 import {withRouter} from "react-router-dom";
 import update from 'immutability-helper';
 import {TextField} from "@material-ui/core";
-import FieldValueGroup from "../../../models/FieldValueGroup";
+import FieldValueGroup from "../../models/FieldValueGroup";
 import Button from "@material-ui/core/Button";
-import AddFieldGroup from "./AddFieldGroup";
+import AddFieldGroup from "./PatientForm/AddFieldGroup";
 import {withSnackbar} from "notistack";
-import FieldGroupsService from "../../../services/FieldGroupsService";
-import FieldValueGroupSection from "./FieldValueGroupForm";
-import FieldValueGroupsService from "../../../services/FieldValueGroupsService";
-import MainContainer from "../../common/MainContainer";
+import FieldGroupsService from "../../services/FieldGroupsService";
+import FieldValueGroupSection from "./PatientForm/FieldValueGroupForm";
+import FieldValueGroupsService from "../../services/FieldValueGroupsService";
+import MainContainer from "../common/MainContainer";
 
 class PatientForm extends Component {
     constructor(props) {
