+++ src/components/Patients/PatientsIndex.jsx
@@ -5,8 +5,8 @@ import PatientsTable from "./PatientsTable";
 import PatientsService from "../../services/PatientsService";
 import {Link} from 'react-router-dom';
 import {Container, Grid, TextField, Typography} from "@material-ui/core";
-import FieldValueGroupSection from "./Forms/FieldValueGroupForm";
-import AddFieldGroup from "./Forms/AddFieldGroup";
+import FieldValueGroupSection from "./PatientForm/FieldValueGroupForm";
+import AddFieldGroup from "./PatientForm/AddFieldGroup";
 import Button from "@material-ui/core/Button";
 import MainContainer from "../common/MainContainer";
